fix: error when creating aux env
This is caused by the fact that we try to fetch the newly created env too quick, so now we retry 10 times waiting 500ms between try.

Also solved a bug when setting the MAX_NUMBER_OF_ENVIRONMENTS and MAX_NUMBER_OF_ALIASES from the env file and not using its default value, it comes as string so when  summing them up it just concatenates them since they come as string.