fix: filter all non-migration files from MIGRATION_DIR
This is useful in our use case because i added a tsc step. So we actually have our migrations as ts files and we compile them each time we run the cmp command. So we have in the scripts key of our package.json:

```
"cmp": "tsc migrations/*.ts && cmp",
```

Could be better, but that's how it is right now. But as a workaround of not supporting ts files from the package itself serves its purpose.